+++ packages/browser/lib/ui/card.ts
@@ -111,6 +111,7 @@ export default class Card {
         hiddenFields: (this.#options.hiddenFields ?? [])?.join(","),
         fields: this.#options.fields,
         acceptedBrands: this.#options.acceptedBrands,
+        customBrands: this.#options.customBrands,
         defaultValues: this.#options.defaultValues,
         autoComplete: this.#options.autoComplete,
         autoProgress: this.#options.autoProgress,