+++ examples/expo/CHANGELOG.md
@@ -1,5 +1,17 @@
 # evervault-expo-example
 
+## 1.0.14
+
+### Patch Changes
+
+- @evervault/react-native@2.6.5
+
+## 1.0.13
+
+### Patch Changes
+
+- @evervault/react-native@2.6.4
+
 ## 1.0.12
 
 ### Patch Changes