+++ examples/react/CHANGELOG.md
@@ -1,5 +1,19 @@
 # example-react
 
+## 0.0.31
+
+### Patch Changes
+
+- Updated dependencies [8501902]
+  - @evervault/react@2.26.0
+
+## 0.0.30
+
+### Patch Changes
+
+- Updated dependencies [52f1154]
+  - @evervault/react@2.25.0
+
 ## 0.0.29
 
 ### Patch Changes