+++ examples/three-d-secure/CHANGELOG.md
@@ -1,5 +1,19 @@
 # example-three-d-secure
 
+## 0.0.51
+
+### Patch Changes
+
+- Updated dependencies [974e08d]
+  - @evervault/js@2.13.0
+
+## 0.0.50
+
+### Patch Changes
+
+- Updated dependencies [52f1154]
+  - @evervault/js@2.12.0
+
 ## 0.0.49
 
 ### Patch Changes